+++ router/pump.go
@@ -303,6 +303,7 @@ func (p *LogsPump) Route(route *Route, logstream chan *Message) {
 		p.mu.Lock()
 		delete(p.routes, updates)
 		p.mu.Unlock()
+		route.closed = true
 	}()
 	for {
 		select {
@@ -371,15 +372,7 @@ func (cp *containerPump) send(msg *Message) {
 		if !route.MatchMessage(msg) {
 			continue
 		}
-		select {
-		case logstream <- msg:
-		case <-time.After(time.Second * 1):
-			debug("pump.send(): send timeout, closing")
-			// normal call to remove() triggered by
-			// route.Closer() may not be able to grab
-			// lock under heavy load, so we delete here
-			defer delete(cp.logstreams, logstream)
-		}
+		logstream <- msg
 	}
 }